+++ call_test.go
@@ -6,17 +6,13 @@ import (
 )
 
 var _ = Describe("call", func() {
-	var testCall Call
-
-	BeforeEach(func() {
-		testCall = &call{
-			args: []interface{}{42, "hello"},
-			out:  []interface{}{40, nil},
-		}
-	})
-
 	Describe("Arguments", func() {
 		It("returns the arguments of the call", func() {
+			testCall := &call{
+				args: []interface{}{42, "hello"},
+				out:  []interface{}{40, nil},
+			}
+
 			result := testCall.Arguments()
 
 			Expect(result).To(Equal([]interface{}{42, "hello"}))
@@ -25,6 +21,11 @@ var _ = Describe("call", func() {
 
 	Describe("ReturnValues", func() {
 		It("returns the return values of the call", func() {
+			testCall := &call{
+				args: []interface{}{42, "hello"},
+				out:  []interface{}{40, nil},
+			}
+
 			result := testCall.ReturnValues()
 
 			Expect(result).To(Equal([]interface{}{40, nil}))