+++ customArguments.go
@@ -18,7 +18,7 @@ func newCustomArguments(stub *mockFunction, arguments []interface{}) *customArgu
 	}
 
 	functionType := stub.toType()
-	if len(arguments) != functionType.NumIn() {
+	if isArgumentLengthValid(functionType, arguments) {
 		return &customArguments{
 			argValidationError: &argumentValidationError{
 				fnType:   functionType,
@@ -36,45 +36,79 @@ func newCustomArguments(stub *mockFunction, arguments []interface{}) *customArgu
 	}
 }
 
+// isArgumentLengthValid returns whether or not the length of the provided arguments
+// are valid for the stub. Taking in variadic function into account.
+func isArgumentLengthValid(functionType reflect.Type, arguments []interface{}) bool {
+	if functionType.IsVariadic() {
+		return len(arguments) < functionType.NumIn()-1
+	}
+
+	return len(arguments) != functionType.NumIn()
+}
+
 // getMatchers returns a slice of matchers based on the types and values of the provided arguments
 func getMatchers(functionType reflect.Type, arguments []interface{}) ([]match.SupportedKindsMatcher, error) {
 	matchers := make([]match.SupportedKindsMatcher, functionType.NumIn())
-	for i, arg := range arguments {
+	for i := 0; i < functionType.NumIn(); i++ {
 		aType := functionType.In(i)
 
-		switch arg.(type) {
-		case match.SupportedKindsMatcher:
-			matcher := arg.(match.SupportedKindsMatcher)
-			if _, ok := matcher.SupportedKinds()[aType.Kind()]; !ok {
-				return nil, &argumentValidationError{
-					fnType:   functionType,
-					provided: arguments,
-				}
+		if isVariadicArgument(functionType, i) {
+			if len(arguments) == functionType.NumIn()-1 {
+				matchers[i] = match.Nil()
+				return matchers, nil
 			}
 
-			matchers[i] = matcher
-		case nil:
-			if !areTypeAndValueEquivalent(aType, arg) {
-				return nil, &argumentValidationError{
-					fnType:   functionType,
-					provided: arguments,
+			variadicArguments := arguments[i:]
+			variadicMatchers := make([]match.SupportedKindsMatcher, len(variadicArguments))
+			for sliceIndex, arg := range variadicArguments {
+				m, found := getMatcher(arg, aType.Elem())
+				if !found {
+					return nil, &argumentValidationError{
+						fnType:   functionType,
+						provided: arguments,
+					}
 				}
+
+				variadicMatchers[sliceIndex] = m
 			}
 
-			matchers[i] = match.Nil()
-		default:
-			if !areTypeAndValueEquivalent(aType, arg) {
-				return nil, &argumentValidationError{
-					fnType:   functionType,
-					provided: arguments,
-				}
-			}
-
-			matchers[i] = match.Exactly(arg)
+			matchers[i] = match.SliceOf(variadicMatchers...)
+			return matchers, nil
 		}
+
+		m, found := getMatcher(arguments[i], aType)
+		if !found {
+			return nil, &argumentValidationError{
+				fnType:   functionType,
+				provided: arguments,
+			}
+		}
+
+		matchers[i] = m
 	}
 
 	return matchers, nil
+}
+
+// getMatcher returns a matcher for the provided type and value
+func getMatcher(value interface{}, valueType reflect.Type) (match.SupportedKindsMatcher, bool) {
+	if matcher, ok := value.(match.SupportedKindsMatcher); ok {
+		if _, ok := matcher.SupportedKinds()[valueType.Kind()]; !ok {
+			return nil, false
+		}
+
+		return matcher, true
+	}
+
+	if !areTypeAndValueEquivalent(valueType, value) {
+		return nil, false
+	}
+
+	if value == nil {
+		return match.Nil(), true
+	}
+
+	return match.Exactly(value), true
 }
 
 type customArguments struct {
@@ -88,12 +122,12 @@ type customArguments struct {
 
 // Return sets the return values for this set of custom arguments
 func (ca *customArguments) Return(returnValues ...interface{}) error {
-	if ca.stub == nil {
-		return errors.New("mocka: stub does not exist")
-	}
-
 	if ca.argValidationError != nil {
 		return ca.argValidationError
+	}
+
+	if ca.stub == nil {
+		return errors.New("mocka: stub does not exist")
 	}
 
 	ca.stub.lock.Lock()