+++ utils.go
@@ -175,3 +175,13 @@ func getType(value interface{}) reflect.Type {
 		return reflect.TypeOf(value)
 	}
 }
+
+// isVariadicArgument returns true if the function is variadic and the argument index
+// is the last argument in the function.
+func isVariadicArgument(functionType reflect.Type, argIndex int) bool {
+	if argIndex == functionType.NumIn()-1 && functionType.IsVariadic() {
+		return functionType.In(argIndex).Kind() == reflect.Slice
+	}
+
+	return false
+}